worker-server/requirements.txt
# Image Processing
Pillow==10.3.0

# Task Queue & Redis
arq==0.25.0
redis==5.0.4

# Networking (For Downloading Logo & Uploading to 8003)
httpx==0.27.0

# Utilities
watchfiles==0.21.0